Move new product stuff and join create stuff to subpackage
