+++ app/src/main/java/kres/realtimeshoppinglist/activities/ListSelectActivity.java
@@ -11,8 +11,8 @@ import android.widget.LinearLayout;
 import java.util.Set;
 
 import kres.realtimeshoppinglist.R;
-import kres.realtimeshoppinglist.dialog.JoinCreateDialog;
-import kres.realtimeshoppinglist.dialog.ShoppingListUtil;
+import kres.realtimeshoppinglist.dialog.joinCreate.JoinCreateDialog;
+import kres.realtimeshoppinglist.dialog.joinCreate.ShoppingListUtil;
 import kres.realtimeshoppinglist.firebase.shoppingList.ListExistsListener;
 import kres.realtimeshoppinglist.firebase.shoppingList.ShoppingListAdapter;
 import kres.realtimeshoppinglist.firebase.shoppingList.ShoppingListManager;
