+++ app/src/main/java/kres/realtimeshoppinglist/dialog/newProduct/NewProductDialog.java
@@ -1,4 +1,4 @@
-package kres.realtimeshoppinglist.dialog;
+package kres.realtimeshoppinglist.dialog.newProduct;
 
 import android.app.Activity;
 import android.app.AlertDialog;
@@ -10,7 +10,6 @@ import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.EditText;
-import android.widget.NumberPicker;
 
 import kres.realtimeshoppinglist.R;
 import kres.realtimeshoppinglist.model.Product;
